docs: Add comprehensive project README, architecture diagrams, and run instructions (#10)
* feat: change amount to integer to represent minor units

- Update database schema from DECIMAL to BIGINT for amount storage
- Modify domain model, DTOs, and entities to use Long type for amount
- Adjust validation logic to work with integer comparisons
- Update test data to use integer amounts representing minor units
- Simplify TransactionResponse DTO to return only essential fields
- Add provider rejection logic for specific amount (999999)
- Update README with comprehensive project documentation

* docs: add design patterns, assumptions, risks, and CI/CD details

Elaborate on applied design patterns, business assumptions, and identified risks. Include details about the CI/CD pipeline, testing strategies, and code quality tools.

+++ src/main/java/com/orchestrator/transaction/application/dto/CreateTransactionCommand.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 import java.time.Instant;
 
 /**
@@ -17,7 +17,7 @@ import java.time.Instant;
 @AllArgsConstructor
 public class CreateTransactionCommand {
     private String clientTransactionId;
-    private BigDecimal amount;
+    private Long amount;
     private String currency;
     private String country;
     private String paymentMethodId;
